January 5, 2012 commit by Barbara Emery.
The TIEGCM modules util.F, wei01gcm.F and wei05sc.F were modified for this 
commit.  calccloc was moved from wei01gcm.F to util.F and modified to use
a default of crad=10deg, offcdeg = 4.2deg, and dskof = 0deg for the 
convection radius and the center location of the convection pattern.
The variables phin and phid used in the Heelis model (phid is also the
MLT location on the dayside of the cusp) were made to be the same as
the default in aurora.F using byloc (byloc is the IMF By in nT but cannot
be less than -11nT or greater than +7nT).  The default was used if the
maximum MLT potential was >12MLT, if the minimum MLT potential was <12MLT,
if dskof<-10deg, if dskof>+10deg, if offcdeg<-5deg, if offcdeg>+10deg,
if crad or crad0 <0deg, and if crad or crad0 >35deg.  Most of the time
the defaults will go into place for Bz northward conditions.

calccloc (CALCulate Convection LOCation) can be called in wei01gcm.F but
was commented out since wei01loc was added to wei01gcm to find crad
from the average BoundaryLat locations from 15.5MLT to 8.5MLT.
(crad = (90-avblat)/2).  The convection center location is offcdeg=4.2
(4.2 deg towards midnight from the magnetic pole) and dskof=0deg (no
offset towards dawn or dusk), which is the same center as for Weimer
2005.  wei01loc is a new routine in wei01gcm.F and is based on the
similar routine wei05loc in wei05sc.F.  Both routines were modified
to use byloc, a subset of byimf, as described above (and used for the
cusp location with phid).

NOTE:  calccloc is NOT CALLED presently by any routine in TIEGCM.
It will be used later for imposed potential patterns like CMIT (LFM)
or AMIE where crad (the convection radius in deg) or theta0 (the
convection radius in radians) is used to define crit(1) in colath.F.
crit(1) and crit(2) define the cross-over co-latitudes of the 
high-latitude and dynamo electric potential patterns, where both are
blended between crit(1) and crit(2).