tiegcm/trunk  - Added namelist parameter JOULEFAC, which has previously been a parameter    set in qjoule.F. This is a convenience for developers, and will not    be included in the default input files.  The default is 1.5.  - Note that if tiegcm is coupled with CMIT, joulefac is hardwired to 1.0    regardless of the value of the namelist parameter (this can easily be    changed if Mike, et.al. want to use the namelist parameter when coupled    with CMIT).  - Wenbin's modifications to filtering for the 2.5-resolution model, as per    his comments below:
Subroutine filter2, which is called by oplus.F and minor.F, has been modified 
to make the 2.5 degree version of the TIEGCM numerically stable. Around the 
two poles the model grid becomes much smaller in the zonal direction. This 
causes numerical instability when plasma and neutral flow speeds are large. 
An FFT filter is thus applied. One surprise occurs near the magnetic equator 
where 2.5-degree TIEGCM is also numerically unstable at the topside of the 
ionosphere. This is most likely caused by the large ambipolar diffusive velocity 
in the model vertical direction that also violates the numerical stability condition. 
Therefore, filtering is also applied in the lower latitudes to ensure the integrity 
of the model solution there.

Two modifications are done so far:

1) An extra parameter 'kutj', which is the latitude index of the grid, is added to the 
   intended in Args.  This allows the filter to know which latitude the fft is applied.

2) An array of kut2 is added to set up the latitude-dependent cutoff wavenumber for the 
   filter. Parameters (O+ density and minor species mixing ratio/density) are strongly 
   filtered near the poles and the equator, less toward the middle latitudes. The filter 
   is also modified so that it does not over filter at high latitudes.

To ensure numerical stability, the model time step has also to be adjusted although 
the FFT filter is applied. For perpetual runs under geomagnetically quiet conditions, 
a model time step of 60 seconds is ok. However, for scientific studies of real 
geophysical conditions, a 30-second time step is recommended. It is also highly 
desirable to check O+ and electron density profiles (especially the topside close 
to the model top boundary) near the poles and the equator to make sure that the model 
solution is good. For geomagnetic storms with Kp large than 6, the model time step 
should be 15 seconds or less.

Wenbin