Rewrite oplus (#21)
* Rewrite oplus

The oplus module is rewritten to split the calculation independent of O+ sub-cycling. This may help speed up the model when there are a lot of O+ sub-cycling.

* Reorganize for clarity

Some post-processing calculations are reorganized for better clarity.

* Code cleanup

Some calculations are rearranged to better reflect the logic. Merge loops and move addfld calls to the end of the subroutine.