Change description of output variable (#36)
* Change output variables and their descriptions

Modified the description of DT_COOL. Change default behavior of saving UI_ExB to include SAPS drifts when SAPS is on. The ion drift without SAPS can be saved separately as UI_wo_SAPS. In addition, the lower bound of SAPS calculation was incorrectly set as Kp=1 in previous versions. Correct it to Kp=0 here.